Umami config & script setup
