+++ src/main/kotlin/sh/zachwal/dailygames/App.kt
@@ -34,7 +34,8 @@ import sh.zachwal.dailygames.guice.HikariModule
 import sh.zachwal.dailygames.guice.JdbiModule
 import sh.zachwal.dailygames.roles.RoleAuthorization
 import sh.zachwal.dailygames.roles.RoleService
-import sh.zachwal.dailygames.sentry.initSentry
+import sh.zachwal.dailygames.config.initSentry
+import sh.zachwal.dailygames.config.initUmami
 import sh.zachwal.dailygames.session.DbSessionStorage
 import sh.zachwal.dailygames.session.SESSION_MAX_DURATION
 import sh.zachwal.dailygames.session.SessionCleanupTask
@@ -63,6 +64,7 @@ fun Application.module(testing: Boolean = false) {
     log.info("Starting app in ${config.env}")
 
     initSentry(config.sentryConfig, config.env)
+    initUmami(config.umamiConfig)
 
     val userService = injector.getInstance(UserService::class.java)
     val roleService = injector.getInstance(RoleService::class.java)
