+++ src/main/kotlin/sh/zachwal/dailygames/config/SentryConfig.kt
@@ -1,6 +1,20 @@
 package sh.zachwal.dailygames.config
 
+import io.sentry.Sentry
+
 data class SentryConfig constructor(
     val kotlinDsn: String,
     val jsDsn: String,
 )
+
+lateinit var jsDsn: String
+lateinit var jsEnv: String
+
+fun initSentry(sentryConfig: SentryConfig, environment: String) {
+    jsDsn = sentryConfig.jsDsn
+    jsEnv = environment
+    Sentry.init { options ->
+        options.dsn = sentryConfig.kotlinDsn
+        options.environment = environment
+    }
+}
