+++ src/main/kotlin/sh/zachwal/dailygames/config/UmamiConfig.kt
@@ -1,0 +1,13 @@
+package sh.zachwal.dailygames.config
+
+data class UmamiConfig(
+    val umamiUrl: String,
+    val websiteId: String,
+)
+
+// Initialize this global variable so it is accessible in HTML utilities
+lateinit var umamiConfig: UmamiConfig
+
+fun initUmami(config: UmamiConfig) {
+    umamiConfig = config
+}
