+++ src/main/kotlin/sh/zachwal/dailygames/shared_html/Head.kt
@@ -5,8 +5,9 @@ import kotlinx.html.link
 import kotlinx.html.meta
 import kotlinx.html.script
 import kotlinx.html.unsafe
-import sh.zachwal.dailygames.sentry.jsDsn
-import sh.zachwal.dailygames.sentry.jsEnv
+import sh.zachwal.dailygames.config.jsDsn
+import sh.zachwal.dailygames.config.jsEnv
+import sh.zachwal.dailygames.config.umamiConfig
 
 fun HEAD.bootstrapCss() {
     link(
@@ -90,6 +91,18 @@ fun HEAD.sentryScript() {
     }
 }
 
+// Points at my personal Umami instance for anonymized site analytics
+// See https://umami.is
+fun HEAD.umamiScript() {
+    if (umamiConfig.umamiUrl.isNotBlank()) {
+        script {
+            defer = true
+            src = "https://${umamiConfig.umamiUrl}/script.js"
+            attributes["data-website-id"] = umamiConfig.websiteId
+        }
+    }
+}
+
 fun HEAD.headSetup() {
     bootstrapCss()
     commonCss()
@@ -97,4 +110,5 @@ fun HEAD.headSetup() {
     favicon()
     mobileUI()
     sentryScript()
+    umamiScript()
 }
