+++ src/main/resources/application.conf
@@ -21,4 +21,10 @@ ktor {
         jsDsn = ""
         jsDsn = ${?SENTRY_JS_DSN}
     }
+    umami {
+       url = ""
+       url = "${?UMAMI_URL}"
+       websiteId = ""
+       websiteId = "${?UMAMI_WEBSITE_ID}"
+    }
 }
